dircmp: test run

+++ src/dircmp/files.py
@@ -156,7 +156,7 @@ def _move_file(path_a : str, path_b : str):
 
 
 
-def execute_operations(oper_list : list[Oper], logger : Callable[[str], None]) -> None:
+def execute_operations(oper_list : list[Oper], logger : Callable[[str], None], test=False) -> None:
     global _break_operations
     _break_operations = False
 
@@ -169,34 +169,34 @@ def execute_operations(oper_list : list[Oper], logger : Callable[[str], None]) -
         try:
             if oper.type == OperType.COPY_AB:
                 logger(f"CP: {oper.path_a} -> {oper.path_b}")
-                _copy_file(oper.path_a, oper.path_b)
+                if not test: _copy_file(oper.path_a, oper.path_b)
                 logger("OK")
             elif oper.type == OperType.COPY_BA:
                 logger(f"CP: {oper.path_b} -> {oper.path_a}")
-                _copy_file(oper.path_b, oper.path_a)
+                if not test: _copy_file(oper.path_b, oper.path_a)
                 logger("OK")
             elif oper.type == OperType.MOVE_AB:
                 logger(f"MV: {oper.path_a} -> {oper.path_b}")
-                _move_file(oper.path_a, oper.path_b)
+                if not test: _move_file(oper.path_a, oper.path_b)
                 logger("OK")
             elif oper.type == OperType.MOVE_BA:
                 logger(f"MV: {oper.path_b} -> {oper.path_a}")
-                _move_file(oper.path_b, oper.path_a)
+                if not test: _move_file(oper.path_b, oper.path_a)
                 logger("OK")
             elif oper.type == OperType.DEL_A:
                 logger(f"RM: {oper.path_a}")
-                Path(oper.path_a).unlink()
+                if not test: Path(oper.path_a).unlink()
                 logger("OK")
             elif oper.type == OperType.DEL_B:
                 logger(f"RM: {oper.path_b}")
-                Path(oper.path_b).unlink()
+                if not test: Path(oper.path_b).unlink()
                 logger("OK")
             elif oper.type == OperType.DEL_AB:
                 logger(f"RM: {oper.path_a}")
-                Path(oper.path_a).unlink()
+                if not test: Path(oper.path_a).unlink()
                 logger("OK")
                 logger(f"RM: {oper.path_b}")
-                Path(oper.path_b).unlink()
+                if not test: Path(oper.path_b).unlink()
                 logger("OK")
         except Exception as e:
             logger("ERROR")
